Add event listener to pages and call stopPropagation on touchmove
This (hopefully) fixes a pesky bug in Safari:

Safari doesn't support CSS's `scroll-behaviour: contain`. When scrolling to the top (or bottom) and waiting for the scroll-bounce to finish , if you then scroll again in the same direction, it completely freezes the scroll (because it scroll bubbles to the parent). But the parent is fixed and non-scrollable, Thus **nothing** happens at all!

This approach hopefully fixes the problem