Browser Optimization - Add global class for styling
Browser Optimization in this case means the following:

Remove the paging mechanism when the page runs in the browser window (not in standalone mode) and stop using position fixed/absolute for the pages.

With Browser Optimizations, the body is the element that scrolls, which means the browser nav bars hide when scrolled (which is otherwise not the case). This is a waste of space, especially on smaller screen and browsers that have a hiding bar at the top and bottom (e.g. Safari on iOS)

Apparently, there is work being done to standardise a way that enables the scrolling of elements within the body (and not the body itself) to trigger this browser-navbar-hiding behavior, but I don't have the url of that article/source right now and it is not supported anywhere yet